requirements.txt
# Azure Functions
azure-functions

# Azure OpenAI SDK
openai>=1.12.0

# Azure Storage (para Blob Trigger)
azure-storage-blob>=12.19.0

# Document processing
python-docx>=1.1.0

# Utilities
python-multipart>=0.0.9

# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry